+++ src/main/java/com/cisowski/schoolmanagement/timetable/attendance/controller/AttendanceController.java
@@ -4,10 +4,7 @@ import com.cisowski.schoolmanagement.common.security.authorization.annotation.Re
 import com.cisowski.schoolmanagement.common.security.authorization.model.ResourceActionType;
 import com.cisowski.schoolmanagement.common.security.authorization.model.ResourceType;
 import com.cisowski.schoolmanagement.common.utility.DbLogger;
-import com.cisowski.schoolmanagement.timetable.attendance.model.AttendanceAbsenceByScheduleResponse;
-import com.cisowski.schoolmanagement.timetable.attendance.model.AttendanceDetailedResponse;
-import com.cisowski.schoolmanagement.timetable.attendance.model.AttendanceSummaryResponse;
-import com.cisowski.schoolmanagement.timetable.attendance.model.MarkAttendanceRequest;
+import com.cisowski.schoolmanagement.timetable.attendance.model.*;
 import com.cisowski.schoolmanagement.timetable.attendance.service.AttendanceService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +12,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 
 @RestController("/attendances")
 @RequiredArgsConstructor
@@ -60,6 +59,16 @@ public class AttendanceController {
     ResponseEntity<List<AttendanceAbsenceByScheduleResponse>> getAbsenceStatsByScheduleForStudent(@PathVariable Integer studentId) {
         DbLogger.info("Received GET request for absence stats for Student with ID: " + studentId);
         List<AttendanceAbsenceByScheduleResponse> response = attendanceService.getAbsenceStatsByScheduleForStudent(studentId);
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+
+    @GetMapping("/student/{studentId}/summary")
+    @RequiresPermission(resource = ResourceType.ATTENDANCE, action = ResourceActionType.READ)
+    ResponseEntity<AttendanceOverallSummaryResponse> getSummaryStatsForStudent(@PathVariable Integer studentId, @RequestParam("startDate") Optional<LocalDate> startDate, @RequestParam("endDate") Optional<LocalDate> endDate) {
+        LocalDate start = startDate.orElse(null);
+        LocalDate end = endDate.orElse(null);
+        DbLogger.info(String.format("Received GET request for summary attendance stats for Student with ID: %s, between %s and %s", studentId, start, end));
+        AttendanceOverallSummaryResponse response = attendanceService.getSummaryAttendanceForStudent(studentId, start, end);
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 }