+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### New
 - ScheduleOccurrence - entity related with Schedule as actual Schedule occurrence within given threshold (created automatically based on coming lessons in X days - changed by Administrator in configuration)
 - Attendance GET endpoints for Teachers
+- Attendance endpoint for summarized Student attendance statistics
 
 ### Changed
 - Creating Attendance - from now attendance is related to ScheduleOccurrence not Schedule itself